# api/openapi.yaml
openapi: 3.0.3
info:
  title: Cerebro API
  description: 'Cerebro is a graph and intelligence platform with security/CSPM as the first
    major application surface. It provides:

    - A shared graph substrate with ontology, provenance, temporal semantics, ingest,
    identity resolution, and simulation primitives

    - Security workflows for policy scanning, findings, compliance, attack-path analysis,
    runtime detection/response, and remediation

    - Org-intelligence workflows built on the same graph substrate

    - Agent-assisted investigation and write-back

'
  version: 2.0.0
  contact:
    name: Security Team
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://api.example.com
    description: Production (replace with your deployment URL)
security:
  - ApiKeyAuth: []
  - BearerAuth: []
...
paths:
...
  /api/v1/query:
    post:
      tags:
        - Assets
      summary: Execute SQL query
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QueryRequest'
      responses:
        '200':
          description: Query results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryResponse'
...
  /api/v1/policies:
    get:
      tags:
        - Policies
      summary: List all policies
      responses:
        '200':
          description: List of policies
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PoliciesResponse'
    post:
      tags:
        - Policies
      summary: Create a new policy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Policy'
      responses:
        '201':
          description: Policy created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Policy'
...
  /api/v1/findings:
    get:
      tags:
        - Findings
      summary: List findings
      parameters:
        - name: severity
          in: query
          schema:
            type: string
            enum:
              - critical
              - high
              - medium
              - low
        - name: status
          in: query
          schema:
            type: string
            enum:
              - open
              - resolved
              - suppressed
        - name: policy_id
          in: query
          schema:
            type: string
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: List of findings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FindingsResponse'
...
  /api/v1/findings/scan:
    post:
      tags:
        - Findings
      summary: Trigger a security scan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                table:
                  type: string
                tables:
                  type: array
                  items:
                    type: string
                limit:
                  type: integer
                  minimum: 1
      responses:
        '200':
          description: Scan results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScanResponse'
        '400':
          description: Invalid scan request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Scan unavailable because Snowflake is not configured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/v1/webhooks:
    get:
      tags:
        - Webhooks
      summary: List webhooks
      responses:
        '200':
          description: List of webhooks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhooksResponse'
    post:
      tags:
        - Webhooks
      summary: Create webhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWebhookRequest'
      responses:
        '201':
          description: Webhook created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Webhook'
...
  /api/v1/audit:
    get:
      tags:
        - Admin
      summary: List audit logs
      description: Query persisted audit entries, optionally filtered by resource.
      parameters:
        - name: resource_type
          in: query
          required: false
          schema:
            type: string
          description: Filter by resource type
        - name: resource_id
          in: query
          required: false
          schema:
            type: string
          description: Filter by resource ID
        - name: limit
          in: query
          required: false
          schema:
            type: integer
          description: Maximum number of log entries to return
      responses:
        '200':
          description: Audit log entries
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '500':
          description: Failed to fetch audit logs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/findings/export:
    get:
      tags:
        - Findings
      summary: Export findings
      description: Export findings in CSV or JSON format with optional filters.
      responses:
        '200':
          description: Exported findings file
          content:
            text/csv:
              schema:
                type: string
                format: binary
            application/json:
              schema:
                type: string
                format: binary
        '500':
          description: Export failure
          content: &id001
            application/json:
              schema:
                type: object
                additionalProperties: true
      parameters:
        - name: severity
          in: query
          required: false
          schema:
            type: string
          description: Filter by severity
        - name: status
          in: query
          required: false
          schema:
            type: string
          description: Filter by status
        - name: policy_id
          in: query
          required: false
          schema:
            type: string
          description: Filter by policy ID
        - name: format
          in: query
          required: false
          schema:
            type: string
            default: csv
            enum:
              - csv
              - json
          description: Export format
        - name: pretty
          in: query
          required: false
          schema:
            type: string
            default: 'false'
            enum:
              - 'true'
              - 'false'
          description: Pretty-print JSON output
...
  /api/v1/impact-analysis:
    post:
      tags:
        - Graph
      summary: Run impact analysis
      description: Analyze cascading impact paths from a starting node under a selected business scenario.
      responses:
        '200':
          description: Impact analysis result
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '400':
          description: Invalid request
          content: *id001
        '503':
          description: Graph unavailable
          content: *id001
      parameters:
        - name: max_depth
          in: query
          required: false
          schema:
            type: integer
            default: 6
          description: Override traversal depth for this request.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - start_node
                - scenario
              properties:
                start_node:
                  type: string
                  description: Starting node ID for traversal.
                scenario:
                  type: string
                  enum:
                    - churn_cascade
                    - revenue_impact
                    - incident_blast_radius
                max_depth:
                  type: integer
                  default: 6
  /api/v1/graph/attack-paths:
    get:
      tags:
        - Graph
      summary: List scored attack paths
      description: Simulate attack paths and return scored results.
      responses:
        '200':
          description: Attack path simulation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttackPathSimulationResponse'
        '503':
          description: Graph unavailable
          content: *id001
      parameters:
        - name: max_depth
          in: query
          required: false
          schema:
            type: integer
            default: 6
          description: Traversal depth (1-10)
        - name: threshold
          in: query
          required: false
          schema:
            type: number
            default: 0
          description: Minimum path score
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
          description: Maximum result size
...
  /api/v1/graph/chokepoints:
    get:
      tags:
        - Graph
      summary: List chokepoints
      description: Return highest-value chokepoints across simulated attack paths.
      responses:
        '200':
          description: Chokepoints
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChokepointsResponse'
        '503':
          description: Graph unavailable
          content: *id001
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
          description: Maximum result size
  /api/v1/graph/compare-permissions:
    get:
      tags:
        - Graph
      summary: Compare permissions
      description: Compare effective permissions between two principals.
      responses:
        '200':
          description: Permission comparison
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PermissionComparisonResponse'
        '400':
          description: principal1/principal2 required
          content: *id001
        '503':
          description: Graph unavailable
          content: *id001
      parameters:
        - name: principal1
          in: query
          required: false
          schema:
            type: string
          description: First principal identifier
        - name: principal2
          in: query
          required: false
          schema:
            type: string
          description: Second principal identifier
...
  /api/v1/graph/stats:
    get:
      tags:
        - Graph
      summary: Get graph statistics
      description: Return graph build metadata and cardinality.
      responses:
        '200':
          description: Graph stats
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GraphStatsResponse'
        '503':
          description: Graph unavailable
          content: *id001
...
  /api/v1/graph/toxic-combinations:
    get:
      tags:
        - Graph
      summary: List toxic combinations
      description: List toxic risk combinations detected in the graph.
      responses:
        '200':
          description: Toxic combinations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToxicCombinationsResponse'
        '503':
          description: Graph unavailable
          content: *id001
      parameters:
        - name: severity
          in: query
          required: false
          schema:
            type: string
          description: Filter by severity
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
          description: Maximum result size
...
  /api/v1/incidents:
    post:
      tags:
        - Runtime
      summary: Create incident
      description: Create an incident record and generate an initial response playbook.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '201':
          description: Incident created
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '400':
          description: Invalid incident payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Failed to create incident
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/v1/providers:
    get:
      tags:
        - Admin
      summary: List providers
      description: List configured providers and sync maturity metadata.
      parameters:
        - name: include_incomplete
          in: query
          required: false
          schema:
            type: boolean
          description: Include providers marked as incomplete
      responses:
        '200':
          description: Provider list
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
...
  /api/v1/rbac/roles:
    get:
      tags:
        - RBAC
      summary: List roles
      description: List RBAC roles.
      responses:
        '200':
          description: Role list
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  additionalProperties: true
        '503':
          description: RBAC unavailable
          content: *id001
  /api/v1/rbac/tenants:
    get:
      tags:
        - RBAC
      summary: List tenants
      description: List RBAC tenants.
      responses:
        '200':
          description: Tenant list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RBACTenant'
        '503':
          description: RBAC unavailable
          content: *id001
    post:
      tags:
        - RBAC
      summary: Create tenant
      description: Create a RBAC tenant.
      responses:
        '201':
          description: Tenant created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RBACTenant'
        '400':
          description: Invalid tenant payload
          content: *id001
        '403':
          description: Permission denied
          content: *id001
        '503':
          description: RBAC unavailable
          content: *id001
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RBACTenantCreateRequest'
  /api/v1/rbac/users:
    post:
      tags:
        - RBAC
      summary: Create user
      description: Create a RBAC user record.
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RBACUser'
        '400':
          description: Invalid user payload
          content: *id001
        '403':
          description: Permission denied
          content: *id001
        '503':
          description: RBAC unavailable
          content: *id001
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RBACUserCreateRequest'
...
  /api/v1/runtime/findings:
    get:
      tags:
        - Runtime
      summary: List runtime findings
      description: List recent runtime findings.
      responses:
        '200':
          description: Runtime findings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuntimeFindingsResponse'
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 100
          description: Maximum result size
...
  /api/v1/telemetry/ingest:
    post:
      tags:
        - Runtime
      summary: Ingest telemetry batch
      description: Ingest a telemetry batch and process each runtime event through detection/response.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TelemetryIngestRequest'
      responses:
        '200':
          description: Telemetry batch processed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TelemetryIngestResponse'
        '400':
          description: Invalid telemetry payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/v1/threatintel/stats:
    get:
      tags:
        - ThreatIntel
      summary: Threat intel stats
      description: Return current threat-intel indicator and feed statistics.
      responses:
        '200':
          description: Threat intel stats
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ThreatIntelStatsResponse'
        '503':
          description: Threat intel service unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/v1/graph/diff:
    get:
      tags:
        - Graph
      summary: Compare graph snapshots by timestamp
      description: Return the structural diff between two graph snapshots identified by RFC3339 timestamps.
      parameters:
        - name: from
          in: query
          required: true
          description: RFC3339 timestamp for the starting snapshot.
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          required: true
          description: RFC3339 timestamp for the ending snapshot.
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Graph diff result
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '400':
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: No snapshots available for the requested timestamps
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Failed to compute graph diff
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
  parameters:
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        default: 50
        maximum: 1000
...
  schemas:
...
    QueryRequest:
      type: object
      required:
        - query
      properties:
        query:
          type: string
          example: SELECT * FROM aws_s3_buckets LIMIT 10
        limit:
          type: integer
    QueryResponse:
      type: object
      properties:
        columns:
          type: array
          items:
            type: string
        rows:
          type: array
          items:
            type: object
        count:
          type: integer
...
    Policy:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        effect:
          type: string
          enum:
            - permit
            - forbid
        resource:
          type: string
        conditions:
          type: array
          items:
            type: string
        severity:
          type: string
          enum:
            - critical
            - high
            - medium
            - low
        tags:
          type: array
          items:
            type: string
    PoliciesResponse:
      type: object
      properties:
        policies:
          type: array
          items:
            $ref: '#/components/schemas/Policy'
        count:
          type: integer
...